# docs/endpoints/openapi-systems-ops.yaml
openapi: 3.0.3
info:
  title: Atlas API — Systems Ops
  version: '1.0.0'
  description: |
    Contrato OpenAPI ULTRA DETALLADO, generado leyendo directamente el código fuente
    (controladores, schemas Zod, mappers, servicios y repositorios), del módulo
    `systems-ops` del backend Atlas: catálogo técnico de endpoints/herramientas/entidades
    de datos, log de acciones (auditoría HTTP), cola de revisión de gobierno de datos,
    perfiles y runs de stress testing, y suites/pasos/runs de pruebas E2E internas.

    ## Envolturas globales (aplican a TODOS los endpoints de este documento)

    Este backend registra dos providers globales en `src/app.module.ts`
    (`APP_INTERCEPTOR` y `APP_FILTER`) que envuelven **toda** respuesta, éxito o error.
    Los schemas de `data` documentados endpoint por endpoint abajo son el contenido
    de la propiedad `data`, **no** el body HTTP completo.

    **Éxito** (`ResponseInterceptor`, `src/common/interceptors/response.interceptor.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "data": { "...": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```

    **Error** (`HttpExceptionFilter`, `src/common/filters/http-exception.filter.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "error": { "code": "VALIDATION_ERROR", "message": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```
    `error.code` se deriva del status HTTP: `400→VALIDATION_ERROR`, `401→UNAUTHORIZED`,
...
    de query/params/body.

...

...
    ```ts
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /systems/dashboard:
    get:
      operationId: getSystemsDashboard
      tags: [SystemsCatalog]
      summary: Vista resumida del estado del catálogo (conteos + postura de gobierno de datos)
      responses:
        '200':
          description: Resumen del dashboard operativo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/endpoints:
    get:
      operationId: listCatalogEndpoints
      tags: [SystemsCatalog]
      summary: Listar endpoints del catálogo con filtros
      parameters:
        - $ref: '#/components/parameters/ModuleQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/RiskLevelQuery'
        - $ref: '#/components/parameters/ReviewStatusQuery'
        - $ref: '#/components/parameters/QQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de endpoints catalogados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_EndpointList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/endpoints/{endpointId}:
    get:
      operationId: getCatalogEndpoint
      tags: [SystemsCatalog]
      summary: Detalle de un endpoint del catálogo (con tools, impactos de datos y de campo)
      parameters:
        - $ref: '#/components/parameters/EndpointIdPath'
      responses:
        '200':
          description: Detalle completo del endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_EndpointDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_ENDPOINT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/endpoints/discover:
    post:
      operationId: discoverEndpoints
      tags: [SystemsCatalog]
      summary: Escanear src/modules/**/*.controller.ts y detectar endpoints automáticamente
      description: Opcionalmente persiste (upsert) cada endpoint detectado y marca como `DEPRECATED_CANDIDATE` los que dejaron de existir en el código.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiscoverEndpointsRequest' }
      responses:
        '201':
          description: Resultado del escaneo (y persistencia si `persist=true`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DiscoverEndpointsResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/endpoints/catalog-seed/refresh:
    post:
      operationId: refreshCatalogSeed
      tags: [SystemsCatalog]
      summary: Re-sembrar el catálogo completo desde fixtures/seeds del backend
      description: Siembra herramientas, entidades de datos (desde modelos Sequelize), endpoints curados + discovery, impactos (desde docs/endpoints/endpoints.md), suites y perfiles de stress. Best-effort — entradas sin match se saltan silenciosamente, no lanza excepciones de negocio.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogSeedRefreshRequest' }
      responses:
        '201':
          description: Conteos de lo sembrado/actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_CatalogSeedRefreshResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/tools:
    get:
      operationId: listTools
      tags: [SystemsCatalog]
      summary: Listar herramientas/dependencias técnicas catalogadas
      parameters:
        - $ref: '#/components/parameters/ModuleQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/RiskLevelQuery'
        - $ref: '#/components/parameters/ReviewStatusQuery'
        - $ref: '#/components/parameters/QQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de herramientas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ToolList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/tools/{toolId}:
    get:
      operationId: getTool
      tags: [SystemsCatalog]
      summary: Detalle de una herramienta catalogada
      parameters:
        - $ref: '#/components/parameters/ToolIdPath'
      responses:
        '200':
          description: Herramienta encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TOOL_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/tools/infer-requirements:
    post:
      operationId: inferToolRequirements
      tags: [SystemsCatalog]
      summary: Inferir qué herramientas usa cada endpoint activo (análisis de patrones sobre el código fuente)
      description: |
        Analiza el código de cada endpoint activo buscando referencias a herramientas conocidas
        (JWT, ZOD, POSTGRES, SEQUELIZE, REDIS, OUTBOX_EVENTS_DB, IDEMPOTENCY_KEYS_DB,
        OPERATIONAL_AUDIT_LOGS, SYSTEM_ACTION_LOGS, S3_OR_OBJECT_STORAGE, WHATSAPP_GENERIC, INFOCENTER)
        y opcionalmente persiste el requisito detectado con `reviewStatus=NEEDS_REVIEW`.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InferToolRequirementsRequest' }
      responses:
        '201':
          description: Resultado de la inferencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_InferToolRequirementsResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/data-entities:
    get:
      operationId: listDataEntities
      tags: [SystemsCatalog]
      summary: Listar entidades de datos (tablas) catalogadas con su clasificación de sensibilidad
      parameters:
        - $ref: '#/components/parameters/ModuleQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/RiskLevelQuery'
        - $ref: '#/components/parameters/ReviewStatusQuery'
        - $ref: '#/components/parameters/QQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de entidades de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataEntityList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/data-entities/{entityId}:
    get:
      operationId: getDataEntity
      tags: [SystemsCatalog]
      summary: Detalle de una entidad de datos
      parameters:
        - $ref: '#/components/parameters/EntityIdPath'
      responses:
        '200':
          description: Entidad encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataEntity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_ENTITY_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/data-entities/{entityId}/metadata:
    patch:
      operationId: updateDataEntityMetadata
      tags: [SystemsCatalog]
      summary: Actualizar metadatos de gobierno de una entidad de datos
      description: |
        **ADVERTENCIA PARA FRONTEND:** este es el ÚNICO endpoint de todo el módulo `systems-ops`
        que NO valida el body con Zod — se recibe como `@Body() body: Record<string, unknown>`
        sin `ZodValidationPipe`. El repositorio (`updateDataEntityMetadata`) aplica una whitelist
        de campos server-side (ver `UpdateDataEntityMetadataRequest` abajo); cualquier otro campo
        enviado se ignora silenciosamente, y no hay control de tipos en runtime más allá de lo
        que Sequelize acepte al persistir (un tipo incompatible puede producir un 500 no controlado
        en vez de un 400 limpio).
      parameters:
        - $ref: '#/components/parameters/EntityIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDataEntityMetadataRequest' }
      responses:
        '200':
          description: Entidad actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DataEntity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_ENTITY_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/impact/by-endpoint/{endpointId}:
    get:
      operationId: getImpactByEndpoint
      tags: [SystemsCatalog]
      summary: Ver, desde un endpoint, qué herramientas usa y qué tablas/campos impacta
      parameters:
        - $ref: '#/components/parameters/EndpointIdPath'
      responses:
        '200':
          description: Impacto del endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ImpactByEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_ENDPOINT_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/impact/by-table/{schemaName}/{tableName}:
    get:
      operationId: getImpactByTable
      tags: [SystemsCatalog]
      summary: Ver, desde una tabla, qué endpoints la impactan
      parameters:
        - name: schemaName
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
          description: Nombre del esquema de BD (p. ej. `public`).
        - name: tableName
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 180 }
          description: Nombre de la tabla.
      responses:
        '200':
          description: Impactos sobre la tabla.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ImpactByTable'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_DATA_ENTITY_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `POST /api/v1/internal/auth/login` (o `/api/v1/auth/login` para
        roles no internos). Debe pertenecer a un usuario con rol en:
        system_admin, platform_admin, admin, qa_engineer, devops, risk_analyst,
        compliance_analyst, readonly_auditor.
  responses:
    BadRequest:
      description: Entrada inválida (Zod) o regla de negocio violada — ver detalle por endpoint.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Falta `Authorization`, formato inválido, token expirado/inválido o revocado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: 'Rol autenticado no está en la lista permitida: "El usuario autenticado no tiene permiso para esta operación."'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  parameters:
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (entero positivo, coercionado desde string).
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Tamaño de página (máx 100).
    ModuleQuery:
      name: module
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Filtra por módulo.
    StatusQuery:
      name: status
      in: query
      schema: { type: string, minLength: 1, maxLength: 40 }
      description: 'Filtra por estado (valores dependen de la entidad, p. ej. ACTIVE, DEPRECATED_CANDIDATE, DISABLED, NEEDS_REVIEW).'
    RiskLevelQuery:
      name: riskLevel
      in: query
      schema: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
      description: Filtra por nivel de riesgo.
    ReviewStatusQuery:
      name: reviewStatus
      in: query
      schema: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
      description: Filtra por estado de revisión.
    QQuery:
      name: q
      in: query
      schema: { type: string, minLength: 1, maxLength: 200 }
      description: Búsqueda de texto libre.
...
    EndpointIdPath:
      name: endpointId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico del endpoint (representado como texto).
    ToolIdPath:
      name: toolId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico de la herramienta.
    EntityIdPath:
      name: entityId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico de la entidad de datos.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        requestId: { type: string, nullable: true, example: 'req-8f3c2a' }
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
            message: { type: string, example: 'SYSTEM_ENDPOINT_NOT_FOUND' }
          required: [code, message]
        timestamp: { type: string, format: date-time }
      required: [error, timestamp]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [page, limit, total, totalPages]

    # ---------- Entidades del dominio (forma real devuelta por los mappers) ----------
    Endpoint:
      type: object
      description: Forma producida por `mapEndpoint` (src/modules/systems-ops/systems-ops.mapper.ts).
      properties:
        endpointId: { type: string }
        code: { type: string }
        module: { type: string }
        controllerName: { type: string, nullable: true }
        handlerName: { type: string, nullable: true }
        method: { type: string }
        routePath: { type: string, description: 'Ruta relativa, sin el prefijo /api/v1.' }
        fullPath: { type: string, description: 'Ruta completa con prefijo /api/v1.' }
        routeName: { type: string }
        businessPurpose: { type: string }
        businessAction: { type: string, nullable: true }
        expectedResponseSummary: { type: string, nullable: true }
        expectedStatusCodes: { type: array, items: { type: integer } }
        minPayloadSchema: { type: object }
        queryParamsSchema: { type: object }
        pathParamsSchema: { type: object }
        headersSchema: { type: object }
        requiresAuth: { type: boolean }
        allowedRoles: { type: array, items: { type: string } }
        containsPii: { type: boolean }
        piiFields: { type: array, items: { type: string } }
        riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        isDestructive: { type: boolean }
        isReadonly: { type: boolean }
        idempotencyRequired: { type: boolean }
        requiresStressTest: { type: boolean }
        requiresIntegrationTest: { type: boolean }
        isTestableFromPortal: { type: boolean }
        testEnvironmentOnly: { type: boolean }
        ownerTeam: { type: string, nullable: true }
        status: { type: string, example: ACTIVE }
        version: { type: string, example: v1 }
        detectedFrom: { type: string, example: controller }
        confidenceLevel: { type: string, example: MEDIUM }
        reviewStatus: { type: string, enum: [AUTO_DETECTED, NEEDS_REVIEW, APPROVED, REJECTED] }
        sourceFile: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
...
    DiscoverEndpointsRequest:
      type: object
      properties:
        mode: { type: string, enum: [SOURCE_SCAN], default: SOURCE_SCAN }
        persist: { type: boolean, default: true }
    CatalogSeedRefreshRequest:
      type: object
      properties:
        includeTools: { type: boolean, default: true }
        includeDataEntities: { type: boolean, default: true }
        includeEndpointSeeds: { type: boolean, default: true }
    InferToolRequirementsRequest:
      type: object
      properties:
        persist: { type: boolean, default: true }
    UpdateDataEntityMetadataRequest:
      type: object
      description: |
        **Sin validación Zod** — cualquier campo fuera de esta whitelist se ignora
        silenciosamente y ningún campo listado aquí se valida por tipo en el pipe
        (solo lo que Sequelize acepte al guardar).
      properties:
        businessPurpose: { type: string }
        dataOwner: { type: string }
        containsPii: { type: boolean }
        containsFinancialData: { type: boolean }
        containsRiskData: { type: boolean }
        containsLegalData: { type: boolean }
        containsDeviceData: { type: boolean }
        containsLocationData: { type: boolean }
        isAuditCritical: { type: boolean }
        retentionPolicyCode: { type: string }
        status: { type: string }
        reviewStatus: { type: string }
...
    DashboardResult:
      type: object
      properties:
        counts:
          type: object
          properties:
            endpoints: { type: integer }
            tools: { type: integer }
            dataEntities: { type: integer }
            testSuites: { type: integer }
            pendingReviews: { type: integer, description: 'Suma de NEEDS_REVIEW en endpoints, entidades, impactos de datos, impactos de campo y requisitos de herramientas.' }
            stressProfiles: { type: integer, description: Perfiles con isEnabled=true. }
            actionLogs24h: { type: integer }
        posture:
          type: object
          properties:
            catalogCoverage: { type: string, enum: [READY_FOR_REVIEW, NEEDS_SEED_REFRESH] }
            pendingReviews: { type: integer }
            stressProfilesEnabled: { type: integer }
...
    Envelope_Dashboard:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DashboardResult' }
    Envelope_EndpointList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Endpoint' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_EndpointDetail:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/EndpointDetailResult' }
...
    Envelope_DiscoverEndpointsResult:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DiscoverEndpointsResult' }
    Envelope_CatalogSeedRefreshResult:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/CatalogSeedRefreshResult' }
    Envelope_ToolList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Tool' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_Tool:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/Tool' }
    Envelope_InferToolRequirementsResult:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/InferToolRequirementsResult' }
    Envelope_DataEntityList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/DataEntity' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_DataEntity:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/DataEntity' }
    Envelope_ImpactByEndpoint:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/ImpactByEndpointResult' }
    Envelope_ImpactByTable:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/ImpactByTableResult' }
...